+++ x/emissions/keeper/msgserver/msg_server_reputer_payload_test.go
@@ -17,7 +17,7 @@ func (s *MsgServerTestSuite) setUpMsgReputerPayload(
 	worker string,
 	workerAddr sdk.AccAddress,
 ) (
-	reputerValueBundle types.ValueBundle,
+	reputerValueBundle types.InputValueBundle,
 	expectedInferences types.Inferences,
 	expectedForecasts types.Forecasts,
 	topicId uint64,
@@ -82,42 +82,49 @@ func (s *MsgServerTestSuite) setUpMsgReputerPayload(
 		},
 	}
 
-	reputerValueBundle = types.ValueBundle{
+	reputerValueBundle = types.InputValueBundle{
 		TopicId:             topicId,
 		ReputerRequestNonce: &types.ReputerRequestNonce{ReputerNonce: &workerNonce},
 		Reputer:             reputerAddr.String(),
 		ExtraData:           nil,
-		CombinedValue:       alloraMath.NewDecFromInt64(100),
-		InfererValues: []*types.WorkerAttributedValue{
+		CombinedValue:       alloraMath.MustNewBoundedExp40Dec(alloraMath.NewDecFromInt64(100)),
+		InfererValues: []*types.InputWorkerAttributedValue{
 			{
 				Worker: workerAddr.String(),
-				Value:  alloraMath.NewDecFromInt64(100),
+				Value:  alloraMath.MustNewBoundedExp40Dec(alloraMath.NewDecFromInt64(100)),
 			},
 		},
-		ForecasterValues: []*types.WorkerAttributedValue{
+		ForecasterValues: []*types.InputWorkerAttributedValue{
 			{
 				Worker: workerAddr.String(),
-				Value:  alloraMath.NewDecFromInt64(100),
+				Value:  alloraMath.MustNewBoundedExp40Dec(alloraMath.NewDecFromInt64(100)),
 			},
 		},
-		NaiveValue:          alloraMath.NewDecFromInt64(100),
-		OneOutInfererValues: []*types.WithheldWorkerAttributedValue{},
-		OneOutForecasterValues: []*types.WithheldWorkerAttributedValue{
+		NaiveValue:          alloraMath.MustNewBoundedExp40Dec(alloraMath.NewDecFromInt64(100)),
+		OneOutInfererValues: []*types.InputWithheldWorkerAttributedValue{},
+		OneOutForecasterValues: []*types.InputWithheldWorkerAttributedValue{
 			{
 				Worker: workerAddr.String(),
-				Value:  alloraMath.NewDecFromInt64(100),
+				Value:  alloraMath.MustNewBoundedExp40Dec(alloraMath.NewDecFromInt64(100)),
 			},
 		},
-		OneInForecasterValues: []*types.WorkerAttributedValue{
+		OneInForecasterValues: []*types.InputWorkerAttributedValue{
 			{
 				Worker: workerAddr.String(),
-				Value:  alloraMath.NewDecFromInt64(100),
+				Value:  alloraMath.MustNewBoundedExp40Dec(alloraMath.NewDecFromInt64(100)),
 			},
 		},
 		OneOutInfererForecasterValues: nil,
 	}
 
 	return reputerValueBundle, expectedInferences, expectedForecasts, topicId
+}
+
+func (s *MsgServerTestSuite) signInputValueBundle(InputReputerValueBundle *types.InputValueBundle, privateKey secp256k1.PrivKey) []byte {
+	require := s.Require()
+	reputerValueBundle, err := types.NewValueBundleFromInput(InputReputerValueBundle)
+	require.NoError(err, "Convert should not return an error")
+	return s.signValueBundle(reputerValueBundle, privateKey)
 }
 
 func (s *MsgServerTestSuite) signValueBundle(reputerValueBundle *types.ValueBundle, privateKey secp256k1.PrivKey) []byte {
@@ -136,15 +143,15 @@ func (s *MsgServerTestSuite) constructAndInsertReputerPayload(
 	reputerAddr sdk.AccAddress,
 	reputerPrivateKey secp256k1.PrivKey,
 	reputerPublicKeyHex string,
-	reputerValueBundle *types.ValueBundle,
+	reputerValueBundle *types.InputValueBundle,
 ) error {
 	ctx, msgServer := s.ctx, s.msgServer
-	valueBundleSignature := s.signValueBundle(reputerValueBundle, reputerPrivateKey)
+	valueBundleSignature := s.signInputValueBundle(reputerValueBundle, reputerPrivateKey)
 
 	// Create a InsertReputerPayloadRequest message
 	lossesMsg := &types.InsertReputerPayloadRequest{
 		Sender: reputerAddr.String(),
-		ReputerValueBundle: &types.ReputerValueBundle{
+		ReputerValueBundle: &types.InputReputerValueBundle{
 			ValueBundle: reputerValueBundle,
 			Signature:   valueBundleSignature,
 			Pubkey:      reputerPublicKeyHex,
@@ -240,12 +247,12 @@ func (s *MsgServerTestSuite) TestMsgInsertReputerPayloadReputerNotMatchSignature
 	s.ctx = sdk.UnwrapSDKContext(s.ctx).WithBlockHeight(newBlockheight)
 
 	reputerValueBundle.Reputer = s.addrsStr[3]
-	valueBundleSignature := s.signValueBundle(&reputerValueBundle, reputerPrivateKey)
+	valueBundleSignature := s.signInputValueBundle(&reputerValueBundle, reputerPrivateKey)
 
 	// Create a InsertReputerPayloadRequest message
 	lossesMsg := &types.InsertReputerPayloadRequest{
 		Sender: reputerAddr.String(),
-		ReputerValueBundle: &types.ReputerValueBundle{
+		ReputerValueBundle: &types.InputReputerValueBundle{
 			ValueBundle: &reputerValueBundle,
 			Signature:   valueBundleSignature,
 			Pubkey:      reputerPublicKeyHex,